refactor(services/*): moved the business rule into the services

+++ src/main/java/com/app/course/services/CourseService.java
@@ -2,10 +2,13 @@ package com.app.course.services;
 
 import java.util.List;
 
+import javax.naming.NameNotFoundException;
+
 import com.app.course.models.CourseEntity;
 import com.app.course.models.CreateCoursePayload;
 import com.app.course.models.UpdateCoursePayload;
 import com.app.course.repositories.CourseRepository;
+import com.app.exceptions.EntityNotFoundException;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -27,19 +30,30 @@ public class CourseService {
     }
 
     public CourseEntity getCourse(String id) {
-        return this.courseRepository.findOne(id);
+        CourseEntity entity = this.courseRepository.findOne(id);
+        if (entity == null)
+            throw new EntityNotFoundException();
+        return entity;
     }
 
     public void deleteCourse(String id) {
+        if (!this.contains(id))
+            throw new EntityNotFoundException();
         this.courseRepository.delete(id);
     }
 
-    public void updateCourse(String id, UpdateCoursePayload coursePayload) {
-        this.courseRepository.update(id, coursePayload);
+    public void updateCourse(String id, UpdateCoursePayload updateCoursePayload) {
+        if (!this.contains(id))
+            throw new EntityNotFoundException();
+        this.courseRepository.update(id, updateCoursePayload);
     }
+
+    //#region Utils
 
     public boolean contains(String id) {
         return this.courseRepository.contains(id);
     }
+
+    //#endregion
 
 }
